Add an `Ord` instance for `VKey`
